Remove deprecated stuff in the Scala.js standard library.
