--- library/src/main/scala/scala/scalajs/concurrent/RunNowExcecutionContext.scala
@@ -1,19 +1,0 @@
-package scala.scalajs.concurrent
-
-import scala.concurrent.ExecutionContextExecutor
-
-private[concurrent] object RunNowExecutionContext
-    extends ExecutionContextExecutor {
-
-  def execute(runnable: Runnable): Unit = {
-    try {
-      runnable.run()
-    } catch {
-      case t: Throwable => reportFailure(t)
-    }
-  }
-
-  def reportFailure(t: Throwable): Unit =
-    t.printStackTrace()
-
-}
