+++ test-suite/js/src/test/scala/org/scalajs/testsuite/jsinterop/MiscInteropTest.scala
@@ -83,7 +83,7 @@ class MiscInteropTest {
      * This will be fixed when we can use ...spread calls with `new`, which
      * we can't yet do because the latest io.js does not support them yet.
      */
-    import scala.scalajs.runtime.assumingES6
+    import scala.scalajs.LinkingInfo.assumingES6
 
     val concreteInstance = {
       val tag = js.constructorTag[ConcreteJSClass]
