+++ test-suite/js/src/test/scala/org/scalajs/testsuite/jsinterop/RuntimeLongTest.scala
@@ -2088,158 +2088,3 @@ class RuntimeLongTest {
   }
 
 }
-
-class RuntimeLongOldTest {
-
-  import RuntimeLong.fromDouble
-
-  def assertHexEquals(expected: String, l: RuntimeLong): Unit =
-    assertEquals(expected, l.toHexString)
-
-  def fromInt(x: Int): RuntimeLong = new RuntimeLong(x)
-
-  val maxInt = fromInt(Int.MaxValue)
-  val minInt = fromInt(Int.MinValue)
-  val one = fromInt(1)
-  val billion = fromInt(1000000000)
-
-  val `4503599627370510L` = new RuntimeLong(14, 0, 256)
-  val `613354684553L` = new RuntimeLong(639113, 146235, 0)
-  val `9863155567412L` = new RuntimeLong(2247476, 2351559, 0)
-  val `3632147899696541255L` = new RuntimeLong(1568327, 2954580, 206463)
-  val `7632147899696541255L` = new RuntimeLong(2616903, 1593290, 433837)
-
-  val minValue = new RuntimeLong(0, 0, 524288)
-  val minus1 = new RuntimeLong(4194303, 4194303, 1048575)
-  val minus2 = new RuntimeLong(4194302, 4194303, 1048575)
-  val minus3 = new RuntimeLong(4194301, 4194303, 1048575)
-  val minus4 = new RuntimeLong(4194300, 4194303, 1048575)
-  val minus15 = new RuntimeLong(4194289, 4194303, 1048575)
-  val minus16 = new RuntimeLong(4194288, 4194303, 1048575)
-
-  @Test def should_correctly_implement_negation(): Unit = {
-    assertHexEquals("fffffffffffffffb", -fromInt(5))
-    assertHexEquals("0", -fromInt(0))
-    assertHexEquals("80000000", -minInt)
-  }
-
-  @Test def should_correctly_implement_comparison(): Unit = {
-    assertEquals(true, fromInt(7)  < fromInt(15))
-    assertEquals(false, fromInt(15) < fromInt(15))
-    assertEquals(true, fromInt(15) <= fromInt(15))
-    assertEquals(true, fromInt(14) <= fromInt(15))
-    assertEquals(false, fromInt(15) > fromInt(15))
-    assertEquals(false, fromInt(14) > fromInt(15))
-    assertEquals(true, fromInt(16) > fromInt(15))
-    assertEquals(true, fromInt(15) >= fromInt(15))
-    assertEquals(false, fromInt(14) >= fromInt(15))
-    assertEquals(true, fromInt(16) >= fromInt(15))
-  }
-
-  @Test def should_correctly_implement_addition(): Unit = {
-    assertHexEquals("16", fromInt(7) + fromInt(15))
-    assertHexEquals("fffffffe", maxInt + maxInt)
-    assertHexEquals("80000000", maxInt + one)
-  }
-
-  @Test def should_correctly_implement_subtraction(): Unit = {
-    assertHexEquals("fffffffffffffff8", fromInt(7) - fromInt(15))
-    assertHexEquals("0", maxInt - maxInt)
-  }
-
-  @Test def should_correctly_implement_multiplication(): Unit = {
-    assertHexEquals("69", fromInt(7)  * fromInt(15))
-    assertHexEquals("ffffffffffffff97", fromInt(-7) * fromInt(15))
-    assertHexEquals("3fffffff00000001", maxInt  * maxInt)
-    assertHexEquals("ffbfffffffffffc8", `4503599627370510L` * fromInt(-4))
-  }
-
-  @Test def should_correctly_implement_division(): Unit = {
-    assertHexEquals("0",  fromInt(7)  / fromInt(15))
-    assertHexEquals("4", fromInt(24) / fromInt(5))
-    assertHexEquals("fffffffffffffffc",  fromInt(24) / fromInt(-5))
-    assertHexEquals("ffffffffe6666667",       maxInt / fromInt(-5))
-    assertHexEquals("2", maxInt / billion)
-    assertHexEquals("2", (maxInt+one) / billion)
-
-    assertHexEquals("1", minValue / minValue)
-    assertHexEquals("8000000000000000", minValue / minus1)
-    assertHexEquals("4000000000000000", minValue / minus2)
-    assertHexEquals("2aaaaaaaaaaaaaaa", minValue / minus3)
-    assertHexEquals("2000000000000000", minValue / minus4)
-    assertHexEquals("888888888888888", minValue / minus15)
-    assertHexEquals("800000000000000", minValue / minus16)
-
-    assertHexEquals("0", `7632147899696541255L` / minValue)
-    assertHexEquals("961529ec0d5811b9", `7632147899696541255L` / minus1)
-    assertHexEquals("cb0a94f606ac08dd", `7632147899696541255L` / minus2)
-    assertHexEquals("dcb1b8a40472b093", `7632147899696541255L` / minus3)
-    assertHexEquals("e5854a7b0356046f", `7632147899696541255L` / minus4)
-    assertHexEquals("f8f05820cdb089b7", `7632147899696541255L` / minus15)
-    assertHexEquals("f961529ec0d5811c", `7632147899696541255L` / minus16)
-  }
-
-  @Test def should_correctly_implement_modulus(): Unit = {
-    assertHexEquals("7", fromInt(7)  % fromInt(15))
-    assertHexEquals("4", fromInt(24) % fromInt(5))
-    assertHexEquals("4", fromInt(24) % fromInt(-5))
-    assertHexEquals("8ca6bff", maxInt % billion)
-    assertHexEquals("8ca6c00", (maxInt+one) % billion)
-    assertHexEquals("2", maxInt % fromInt(-5))
-
-    assertHexEquals("0", minValue % minValue)
-    assertHexEquals("0", minValue % minus1)
-    assertHexEquals("0", minValue % minus2)
-    assertHexEquals("fffffffffffffffe", minValue % minus3)
-    assertHexEquals("0", minValue % minus4)
-    assertHexEquals("fffffffffffffff8", minValue % minus15)
-    assertHexEquals("0", minValue % minus16)
-
-    assertHexEquals("69ead613f2a7ee47", `7632147899696541255L` % minValue)
-    assertHexEquals("0", `7632147899696541255L` % minus1)
-    assertHexEquals("1", `7632147899696541255L` % minus2)
-    assertHexEquals("0", `7632147899696541255L` % minus3)
-    assertHexEquals("3", `7632147899696541255L` % minus4)
-    assertHexEquals("0", `7632147899696541255L` % minus15)
-    assertHexEquals("7", `7632147899696541255L` % minus16)
-  }
-
-  @Test def should_correctly_implement_toString(): Unit = {
-    assertEquals("2147483647", maxInt.toString)
-    assertEquals("-50", fromInt(-50).toString)
-    assertEquals("-1000000000", fromInt(-1000000000).toString)
-    assertEquals("2147483648", (maxInt+one).toString)
-    assertEquals("-2147483648", minInt.toString)
-  }
-
-  @Test def should_correctly_implement_fromDouble(): Unit = {
-    assertHexEquals("4", fromDouble(4.5))
-    assertHexEquals("fffffffffffffffc", fromDouble(-4.5))
-  }
-
-  @Test def should_correctly_implement_toDouble(): Unit = {
-    assertEquals(5.0, fromInt(5).toDouble, 0.0)
-    assertEquals(2147483648.0, (maxInt+one).toDouble, 0.0)
-  }
-
-  @Test def should_correctly_implement_numberOfLeadingZeros(): Unit = {
-    assertEquals(64, fromInt(0).numberOfLeadingZeros)
-    assertEquals(63, fromInt(1).numberOfLeadingZeros)
-    assertEquals(0, fromInt(-1).numberOfLeadingZeros)
-    assertEquals(62, fromInt(2).numberOfLeadingZeros)
-  }
-
-  @Test def should_implement_hashCode_according_to_spec_in_j_l_Long(): Unit = {
-    assertEquals(0, fromInt(0).hashCode())
-    assertEquals(55, fromInt(55).hashCode())
-    assertEquals(11, fromInt(-12).hashCode())
-    assertEquals(10006548, fromInt(10006548).hashCode())
-    assertEquals(1098747, fromInt(-1098748).hashCode())
-
-    assertEquals(-825638905, `613354684553L`.hashCode())
-    assertEquals(1910653900, `9863155567412L`.hashCode())
-    assertEquals(1735398658, `3632147899696541255L`.hashCode())
-    assertEquals(-1689438124, `7632147899696541255L`.hashCode())
-  }
-
-}
